fixing bugs + refactor + new tests

+++ src/main/java/org/pl/deenes/api/controller/rest/TrainRestController.java
@@ -5,9 +5,12 @@ import lombok.AllArgsConstructor;
 import org.pl.deenes.api.controller.dto.TrainDTO;
 import org.pl.deenes.api.controller.mapper.TrainDTOMapper;
 import org.pl.deenes.infrastructure.repositories.TrainRepository;
+import org.pl.deenes.model.Train;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @AllArgsConstructor
@@ -18,16 +21,26 @@ public class TrainRestController {
     private final TrainRepository trainRepository;
 
     @GetMapping(value = "/{trainKwr}")
-    public List<TrainDTO> trainInfo(@PathVariable Integer trainKwr) {
-
-        var train = trainRepository.find(trainKwr);
-        //.orElseThrow(() -> new NotFound("Train kwr: %s not found".formatted(trainKwr)));
-        return train.stream().map(trainDTOMapper::mapToDTO).toList();
+    public ResponseEntity<List<TrainDTO>> trainInfo(@PathVariable Integer trainKwr) {
+        var trainOptional = trainRepository.find(trainKwr);
+        if (trainOptional.isPresent()) {
+            List<TrainDTO> trainDTOList = trainOptional.stream()
+                    .map(trainDTOMapper::mapToDTO)
+                    .toList();
+            return ResponseEntity.ok(trainDTOList);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping(value = "/delete/{trainKwr}")
-    public void deleteTrain(@PathVariable Integer trainKwr) {
-        trainRepository.delete(trainKwr);
+    public ResponseEntity<String> deleteTrain(@PathVariable Integer trainKwr) {
+        Optional<Train> trainOptional = trainRepository.find(trainKwr);
+        if (trainOptional.isPresent()) {
+            trainRepository.delete(trainKwr);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
-
 }
